AOSP security patch correction
Below commit wrongly applied the AOSP patch to NXP code base.
Instead of modifying tRW_T2T_LOCK_INFO struct, tRW_T1T_LOCK_INFO struct was updated.

https://www.collabnet.nxp.com/gerrit/c/nxp_nci_hal_libnfc-nci/+/92072

+++ SN100x/src/nfc/include/rw_int.h
@@ -154,7 +154,7 @@ typedef uint8_t tRW_T1T_LOCK_STATUS;
 typedef struct {
   uint16_t offset;  /* Offset of the lock byte in the Tag                   */
   uint8_t num_bits; /* Number of lock bits in the lock byte                 */
-  uint16_t bytes_locked_per_bit; /* No. of tag bytes gets locked by a bit in this
+  uint8_t bytes_locked_per_bit; /* No. of tag bytes gets locked by a bit in this
                                    byte   */
 } tRW_T1T_LOCK_INFO;
 
@@ -385,7 +385,7 @@ typedef uint8_t tRW_T2T_LOCK_STATUS;
 typedef struct {
   uint16_t offset;              /* Offset of the lock byte in the Tag */
   uint16_t num_bits;            /* Number of lock bits in the lock byte */
-  uint8_t bytes_locked_per_bit; /* No. of tag bytes gets locked by a bit in this
+  uint16_t bytes_locked_per_bit; /* No. of tag bytes gets locked by a bit in this
                                    byte       */
 } tRW_T2T_LOCK_INFO;
